Engine to recommend stocks depending upon the strategy
Engine developed in Python which suggests which Stocks to buy depending
on the strategies selected by users. UI is developed in HTML5, JS and
CSS. 5 strategies are suggested to the user. Some default companies are
in each category. Uses Google and Yahoo finance services to get latest
stock data and to get past history of the company stock.